// api/src/enums/errors.enum.ts
export enum Errors {
	WEAK_PASSWORD = "WEAK_PASSWORD",
	PASSWORD_MISMATCH = "PASSWORD_MISMATCH",
	INVALID_PASSWORD = "INVALID_PASSWORD",
	INTERNAL_ERROR = "INTERNAL_ERROR",
	INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
	INVALID_TOKEN = "INVALID_TOKEN",
	USER_NOT_FOUND = "USER_NOT_FOUND",
	USER_ALREADY_EXISTS = "USER_ALREADY_EXISTS",
	ALREADY_EXISTS = "ALREADY_EXISTS",
	EMAIL_ALREADY_TAKEN = "EMAIL_ALREADY_TAKEN",
	INVALID_REQUEST_DATA = "INVALID_REQUEST_DATA",
	INVALID_PARAMETER = "INVALID_PARAMETER",
	NOT_FOUND = "NOT_FOUND",
	BATCH_NOT_FOUND = "BATCH_NOT_FOUND",
	TRANSACTION_NOT_FOUND = "TRANSACTION_NOT_FOUND",
	EMAIL_NOT_SENT = "EMAIL_NOT_SENT",
	INVALID_OTP = "INVALID_OTP",
	OTP_MAX_ATTEMPTS = "OTP_MAX_ATTEMPTS",
	ADDRESS_MISSING = "ADDRESS_MISSING",
	IMAGE_CONVERSION = "IMAGE_CONVERSION",
	IMAGE_UPLOAD = "IMAGE_UPLOAD",
	STRIPE_ERROR = "STRIPE_ERROR",
	JWT_ALREADY_BLACKLISTED = "JWT_ALREADY_BLACKLISTED",
	INVALID_OTP_AUTH_CODE = "INVALID_OTP_AUTH_CODE",
	FORBIDDEN = "FORBIDDEN",
	CUSTOM_COLUMN_ALREADY_EXISTS = "CUSTOM_COLUMN_ALREADY_EXISTS",
	ASSIGNMENT_FAILED = "ASSIGNMENT_FAILED",
	NOT_ENOUGH_CREDITS = "NOT_ENOUGH_CREDITS",
}